Rewamp settings screen
