+++ lib/widgets/home.dart
@@ -26,7 +26,7 @@ import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 
 import 'package:bluecherry_client/widgets/device_grid/device_grid.dart';
 import 'package:bluecherry_client/widgets/events_screen.dart';
-import 'package:bluecherry_client/widgets/settings.dart';
+import 'package:bluecherry_client/widgets/settings/settings.dart';
 import 'package:bluecherry_client/utils/methods.dart';
 import 'package:bluecherry_client/widgets/add_server_wizard.dart';
 import 'package:bluecherry_client/widgets/direct_camera.dart';
